+++ packages/elements-react/src/context/component.tsx
@@ -47,6 +47,7 @@ export function useGroupSorter() {
 
 const defaultNodeOrder = [
   "oidc",
+  "saml",
   "identifier_first",
   "default",
   "profile",