+++ packages/elements-react/src/theme/default/assets/types.ts
@@ -1,0 +1,6 @@
+// Copyright © 2025 Ory Corp
+// SPDX-License-Identifier: Apache-2.0
+
+export type SVGIcon = React.FunctionComponent<
+  React.ComponentProps<"svg"> & { size?: number }
+>